Moved foot_U to Syntax, and removed StockFlowUnitsSyntax
